Finish packaging tictactoe
